Few improvements
